deleted unneccessary components and pages, combined login and register as dialogs onto home page
