+++ client/src/App.jsx
@@ -3,12 +3,9 @@ import { Router, Route } from 'react-router-dom';
 import Navbar from './components/Navbar/Navbar.jsx';
 import Home from './pages/home.jsx';
 import dashboardTeacher from './pages/dashboardTeacher.jsx';
-import dashboardStudent from './pages/dashboardStudent'
 import Grades from './pages/gradesTeacher.jsx';
-import Login from './pages/login.jsx';
 import Logout from './pages/logout.jsx';
 import Search from './pages/search.jsx';
-// import Register from './components/RegisterForm/Register.jsx';
 import Classroom from './pages/Classroom.jsx';
 import Footer from './components/Footer/Footer.jsx';
 import Wrapper from './components/Wrapper/Wrapper.jsx';
@@ -41,15 +38,10 @@ const App = () => {
                     <React.Fragment>
                         <Route exact path="/search" component={Search} />
                         <Route exact path="/dashboardTeacher" component={dashboardTeacher} />
-                        <Route exact path="/dashboardStudent" component={dashboardStudent} />
                         <Route exact path="/grades" component={Grades} />
-                        {/* <Route exact path="/register" component={Register} /> */}
                         <Route exact path="/classrooms" component={Classroom} />
                         <Route exact path="/" component={Home} />
                         <Route exact path="/logout" component={Logout} />
-                        <Route path="/login">
-                            <Login/>
-                        </Route>
                     </React.Fragment>
                 </Wrapper>
                 <Footer />
