--- client/src/components/Assignments/TeacherDashboard.jsx
@@ -1,147 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
-import ButtonGroup from '@material-ui/core/ButtonGroup';
-import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
-import ClickAwayListener from '@material-ui/core/ClickAwayListener';
-import Grow from '@material-ui/core/Grow';
-import Popper from '@material-ui/core/Popper';
-import MenuItem from '@material-ui/core/MenuItem';
-import MenuList from '@material-ui/core/MenuList';
-import TextField from '@material-ui/core/TextField';
-
-const options = ['Select a class', 'Class title1', 'Class title2', 'Class title 3'];
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-  },
-  // eslint-disable-next-line
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
- 
-    },
-  },
-}));
-
-export default function TeachDashboard() {
-  const classes = useStyles();
-  const [open, setOpen] = React.useState(false);
-  const anchorRef = React.useRef(null);
-  const [selectedIndex, setSelectedIndex] = React.useState(1);
-
-  const handleClick = () => {
-    console.info(`You clicked ${options[selectedIndex]}`);
-  };
-
-  const handleMenuItemClick = (event, index) => {
-    setSelectedIndex(index);
-    setOpen(false);
-  };
-
-  const handleToggle = () => {
-    setOpen((prevOpen) => !prevOpen);
-  };
-
-  const handleClose = (event) => {
-    if (anchorRef.current && anchorRef.current.contains(event.target)) {
-      return;
-    }
-
-    setOpen(false);
-};
-  return (
-    <div className= {classes.root}>
-
-      <Grid container
-      spacing={3}
-      direction="row"
-      justify="center"
-      alignItems="stretch">
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>
-          <h5>Add an assignment:</h5>
-              <form className={classes.root} noValidate autoComplete="off">
-                  <TextField fullWidth id="standard-basic" label="Assignment Title" />
-                  <TextField fullWidth id="standard-basic" label="Assignment Due Date" />
-                  <TextField fullWidth
-                    id="outlined-multiline-static"
-                    label="Description:"
-                    multiline
-                    rows="8"
-                    defaultValue="Add the description of the assignment here"
-                    variant="outlined"
-                  />
-                </form>
-          </Paper>
-        </Grid>
-        <Grid item xs={6}>
-          <Paper className={classes.paper}>
-          <h5>Preview assignment:</h5>
-         
-          </Paper>
-        </Grid>
-
-        <Grid item xs={12}>
-          <Paper className={classes.paper}>
-            <ButtonGroup variant="contained" color="primary" ref={anchorRef} aria-label="split button">
-                <Button onClick={handleClick}>{options[selectedIndex]}</Button>
-                    <Button
-                      color="primary"
-                      size="small"
-                      aria-controls={open ? 'split-button-menu' : undefined}
-                      aria-expanded={open ? 'true' : undefined}
-                      aria-label="select merge strategy"
-                      aria-haspopup="menu"
-                      onClick={handleToggle}
-                    >
-                      <ArrowDropDownIcon />
-                    </Button>
-            </ButtonGroup>
-            
-            <Popper open={open} anchorEl={anchorRef.current} role={undefined} transition disablePortal>
-              {({ TransitionProps, placement }) => (
-                <Grow
-                  {...TransitionProps}
-                  style={{
-                    transformOrigin: placement === 'bottom' ? 'center top' : 'center bottom',
-                  }}
-                >
-                  <Paper>
-                    <ClickAwayListener onClickAway={handleClose}>
-                      <MenuList id="split-button-menu">
-                        {options.map((option, index) => (
-                          <MenuItem
-                            key={option}
-                            disabled={index === 2}
-                            selected={index === selectedIndex}
-                            onClick={(event) => handleMenuItemClick(event, index)}
-                          >
-                            {option}
-                          </MenuItem>
-                        ))}
-                      </MenuList>
-                    </ClickAwayListener>
-                  </Paper>
-                </Grow>
-              )}
-            </Popper>
-            <h8>&nbsp;&nbsp;&nbsp;</h8>
-            <Button variant="contained" color="primary"> 
-                  SUBMIT
-            </Button>
-          </Paper>
-        </Grid>
-      </Grid>  
-    </div>
-    
-  );
-}
