+++ client/src/components/LoginForm/style.css
@@ -1,0 +1,3 @@
+.dialog-class {
+    background-color: red !important;
+}
