+++ client/src/pages/Classroom.jsx
@@ -59,7 +59,7 @@ export const Classroom = (props) => {
         }
         catch (error) {
             console.log(error)
-            history.replace('/login')
+            history.replace('/')
         }
     }
